requirements/production.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --output-file=requirements/production.txt requirements/production.in
#
adjusttext==0.7.3
    # via -r requirements/./base.in
argon2-cffi==21.3.0
    # via -r requirements/./base.in
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
asgiref==3.6.0
    # via django
auth0-python==3.24.0
    # via -r requirements/./base.in
blinker==1.5
    # via nplusone
boto3==1.26.43
    # via
    #   django-storages
    #   ssm-parameter-store
botocore==1.29.43
    # via
    #   boto3
    #   s3transfer
brotli==1.0.9
    # via
    #   fonttools
    #   whitenoise
certifi==2022.12.7
    # via
    #   requests
    #   sentry-sdk
cffi==1.15.1
    # via
    #   argon2-cffi-bindings
    #   cryptography
    #   weasyprint
charset-normalizer==2.1.1
    # via requests
collectfast==2.2.0
    # via -r requirements/production.in
contourpy==1.0.6
    # via matplotlib
coreapi==2.3.3
    # via -r requirements/./base.in
coreschema==0.0.4
    # via coreapi
cryptography==39.0.0
    # via
    #   pyjwt
    #   social-auth-core
cssselect2==0.7.0
    # via weasyprint
cycler==0.11.0
    # via matplotlib
defusedxml==0.7.1
    # via
    #   odfpy
    #   python3-openid
    #   social-auth-core
diff-match-patch==20200713
    # via django-import-export
django==4.1.5
    # via
    #   -r requirements/./base.in
    #   collectfast
    #   django-anymail
    #   django-cors-headers
    #   django-import-export
    #   django-model-utils
    #   django-storages
    #   djangorestframework
django-anymail[mailgun]==9.0
    # via -r requirements/production.in
django-cors-headers==3.13.0
    # via -r requirements/./base.in
django-crispy-forms==1.14.0
    # via -r requirements/./base.in
django-environ==0.9.0
    # via -r requirements/./base.in
django-import-export==3.0.2
    # via -r requirements/./base.in
django-model-utils==4.3.1
    # via -r requirements/./base.in
django-storages[boto3]==1.13.2
    # via
    #   -r requirements/production.in
    #   collectfast
django-waffle==3.0.0
    # via -r requirements/./base.in
djangorestframework==3.14.0
    # via -r requirements/./base.in
ecdsa==0.18.0
    # via python-jose
et-xmlfile==1.1.0
    # via openpyxl
fonttools[woff]==4.38.0
    # via
    #   matplotlib
    #   weasyprint
gunicorn==20.1.0
    # via -r requirements/production.in
html5lib==1.1
    # via weasyprint
idna==3.4
    # via requests
itypes==1.2.0
    # via coreapi
jinja2==3.1.2
    # via
    #   -r requirements/./base.in
    #   coreschema
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
kiwisolver==1.4.4
    # via matplotlib
markuppy==1.14
    # via tablib
markupsafe==2.1.1
    # via jinja2
matplotlib==3.6.2
    # via
    #   -r requirements/./base.in
    #   adjusttext
nplusone==1.0.0
    # via -r requirements/./base.in
numpy==1.24.1
    # via
    #   adjusttext
    #   contourpy
    #   matplotlib
oauthlib==3.2.2
    # via
    #   requests-oauthlib
    #   social-auth-core
odfpy==1.4.1
    # via tablib
openpyxl==3.0.10
    # via tablib
packaging==22.0
    # via matplotlib
pillow==9.4.0
    # via
    #   -r requirements/./base.in
    #   matplotlib
    #   weasyprint
psycopg2-binary==2.9.5
    # via -r requirements/./base.in
pyasn1==0.4.8
    # via
    #   python-jose
    #   rsa
pycparser==2.21
    # via cffi
pydantic==1.10.4
    # via -r requirements/./base.in
pydyf==0.5.0
    # via weasyprint
pyjwt[crypto]==2.6.0
    # via
    #   auth0-python
    #   social-auth-core
pyparsing==3.0.9
    # via matplotlib
pyphen==0.13.2
    # via weasyprint
python-dateutil==2.8.2
    # via
    #   botocore
    #   matplotlib
python-jose==3.3.0
    # via -r requirements/./base.in
python3-openid==3.2.0
    # via social-auth-core
pytz==2022.7
    # via djangorestframework
pyyaml==6.0
    # via tablib
requests==2.28.1
    # via
    #   -r requirements/./base.in
    #   auth0-python
    #   coreapi
    #   django-anymail
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==1.3.1
    # via social-auth-core
returns==0.19.0
    # via -r requirements/./base.in
rsa==4.9
    # via python-jose
s3transfer==0.6.0
    # via boto3
sentry-sdk==1.12.1
    # via -r requirements/./base.in
six==1.16.0
    # via
    #   ecdsa
    #   html5lib
    #   nplusone
    #   python-dateutil
social-auth-app-django==5.0.0
    # via -r requirements/./base.in
social-auth-core==4.3.0
    # via social-auth-app-django
sqlparse==0.4.3
    # via django
ssm-parameter-store==19.11.0
    # via -r requirements/./base.in
tablib[html,ods,xls,xlsx,yaml]==3.3.0
    # via django-import-export
tinycss2==1.2.1
    # via
    #   cssselect2
    #   weasyprint
typing-extensions==4.4.0
    # via
    #   collectfast
    #   pydantic
    #   returns
uritemplate==4.1.1
    # via coreapi
urllib3==1.26.13
    # via
    #   botocore
    #   requests
    #   sentry-sdk
weasyprint==57.2
    # via -r requirements/./base.in
webencodings==0.5.1
    # via
    #   cssselect2
    #   html5lib
    #   tinycss2
whitenoise[brotli]==6.3.0
    # via -r requirements/./base.in
xlrd==2.0.1
    # via tablib
xlwt==1.3.0
    # via tablib
zopfli==0.2.2
    # via fonttools

# The following packages are considered to be unsafe in a requirements file:
# setuptools
